TravelPlannerApp - django backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-rest-swagger==2.2.0
djangorestframework
djangorestframework-dataclasses==1.2.0
djangorestframework-simplejwt==5.2.2
Faker
faker-vehicle==0.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
psycopg2==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
simplejson==3.18.4
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
